requirements.txt
fastapi
uvicorn
asyncpg
passlib[bcrypt]
python-jose[cryptography]
python-dotenv